personal-finance-chatbot/frontend/streamlit/requirements.txt
streamlit==1.24.1
requests==2.31.0
python-dotenv